Reference dev.yml's 'orphan' environment variables from dev.env
This is to elaborate on the proposal made in #1082. 

Contain 'orphan' variables scattered all around dev.yml in a separate dev.env file which can be stored in VCS likewise the development-specific variables it defines are.

N.B. `django` service's `USE_DOCKER=yes` variable left intact for its use to be isolated within the service only.