Add failing test for travis.yml
I see three options to test travis.yml :

1. Testing that the YAML contains relevant value. Least useful and least
reliable, but simplest to implement.

2. Testing that the YAML is valid TravisCI YAML. Unfortunately this is
difficult / impossible. Doing 'travis lint' would succeed, this command
does not check for 'script' key presence and wouldn't be useful for us.

We could use 'travis-build' to verify that the YAML can be converted to
a worker config, but as of now 'travis-build' doesn't work out of the
box.

There is a new tool for validating travis YAML files 'travis-yml', but
as of now it's a ruby-only library and it's still a work in progress.

3. Running Travis CI task based on the generated YAML. This seems the
best approach, however since cookiecutter-django itself uses Travis CI,
that would require running Travis CI from within Travis CI.

Scheduling Travis CI job without a github push still requires a public
github repo, which is something that we can't generate on demand.

Given that I'm opting to use approach 1.