Adds missing config to generated .travis.yml
The keys added are as follows:

1. 'script'

Required by Travis, cookiecutter-django used to provide it until it has
been removed together with hitch.

I'm assuming hitch has been replaced with pytest, I'm setting pytest as
the new value for the 'script' key.

2. 'install'

Not required by Travis, but necessary in our case; installs test
libraries, mostly pytest.

As of now this points to 'local.txt' requirements file. There used to be
a separate 'test.txt' requirements file but it has been decided to merge
it with 'local.txt', see discussion in
https://github.com/pydanny/cookiecutter-django/pull/1557 .