Fix APPS path when using pytest
Pytest does not use manage.py when executing tests. This means that the
trick we use to append app directories inside of the main project won't
work since the logic exists in manage.py.

If we instead place the same logic in base.py, things work as expected
when running pytest.