Add django-admin startapp template
django-admin has a --template flag that lets you specify a directory to
base new apps off of. I've added a `startapp_template` directory to the
top level directory, updated the post gen hook to remove any files that
may not be needed, and updated the documentation for this new process.